add toolbar button and modal to display the tree for DRP projects
