+++ client/src/components/DataFiles/DataFilesModals/DataFilesModals.jsx
@@ -19,6 +19,7 @@ import DataFilesDownloadMessageModal from './DataFilesDownloadMessageModal';
 import './DataFilesModals.scss';
 import DataFilesFormModal from './DataFilesFormModal';
 import DataFilesPublicationRequestModal from './DataFilesPublicationRequestModal';
+import DataFilesProjectTreeModal from './DataFilesProjectTreeModal';
 
 export default function DataFilesModals() {
   return (
@@ -41,6 +42,7 @@ export default function DataFilesModals() {
       <DataFilesMakePublicModal />
       <DataFilesDownloadMessageModal />
       <DataFilesFormModal />
+      <DataFilesProjectTreeModal/>
       <DataFilesPublicationRequestModal />
     </>
   );
