+++ client/vite.config.ts
@@ -19,6 +19,7 @@ export default defineConfig({
   resolve: {
     alias: {
       _common: resolve(__dirname, 'src/components/_common'),
+      _custom: resolve(__dirname, 'src/components/_custom'),
       hooks: resolve(__dirname, 'src/hooks'),
       utils: resolve(__dirname, 'src/utils'),
       styles: resolve(__dirname, 'src/styles'),
